guaterpura/requirements.txt
cairocffi==1.1.0
CairoSVG==2.4.2
cffi==1.14.0
cssselect2==0.3.0
defusedxml==0.6.0
Django==2.2.7
django-easy-pdf==0.1.1
html5lib==1.0.1
Pillow==7.1.2
psycopg2==2.8.5
pycparser==2.20
PyPDF2==1.26.0
Pyphen==0.9.5
pytz==2020.1
reportlab==3.5.42
six==1.15.0
sqlparse==0.3.1
tinycss2==1.0.2
WeasyPrint==51
webencodings==0.5.1
xhtml2pdf==0.2.4
